catalog-service/requirements.txt
flask
flask-restful
cassandra-driver
